safe-harbor/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
python-multipart==0.0.18
websockets==14.*
openpyxl==3.1.*
pydantic==2.10.*
pydantic-settings==2.*
openai==1.*
google-genai==0.3.*
python-dotenv==1.0.*
aiofiles==24.*
